requirements.txt
functions-framework==3.*
google-api-python-client
google-cloud-bigquery
google-cloud-storage
pandas
requests
openpyxl
pyarrow